Modif. correcciones de codigo y comentado el caso de recurrencia + Java Doc v 1.0

+++ src/Caso1/Iteracion.java
@@ -1,21 +1,23 @@
 package Caso1;
 /**
  * <h1>Aplicacion del metodo iterativo.</h1>
+ * 
+ * <p>
+ * 		Se crea una variable "acumulador" que almacena la suma sucesiva del sumatorio.
+ * 		Se crea una variable "resultado" que almacena el valor final del sumatorio.
+ * 
+ * 		El for itera las veces necesarias, sustituyendo el valor de acumulador.
+ * 		Al finalizar, el valor de acumulador se asigna a la variable "resultado".
+ * </p>
+ * 
  * @author German P.
- * 
- * Se crea una variable "acumulador" que almacena la suma sucesiva del sumatorio.
- * Se crea una variable "resultado" que almacena el valor final del sumatorio.
- * 
- * El for itera las veces necesarias, sustituyendo el valor de acumulador.
- * Al finalizar, el valor de acumulador se asigna a la variable "resultado".
- *
  */
 
 public class Iteracion {
-	public static int iteracion(int valorEntrada) {
+	public static double iteracion(int valorEntrada) {
 
         int acumulador = 0;	//Acumula los resultados del sumatorio
-        int resultado = 0;		//Contendra el resultado final
+        double resultado = 0;		//Contendra el resultado final
 
         for (int i = 0 ; i <= (valorEntrada-1); i++) {
             acumulador += (4 * i) + 1;
